- created typedict for speechconfig - simplified gemini tts model detection - moved gemini_tts test to test_litellm
